[utils] move dialog box to file
